+++ networking/eth2/src/testFixtures/java/tech/pegasys/teku/networking/eth2/Eth2P2PNetworkFactory.java
@@ -291,9 +291,7 @@ public class Eth2P2PNetworkFactory {
 
         if (rpcEncoding == null) {
           rpcEncoding =
-              RpcEncoding.createSszSnappyEncoding(
-                  spec.getNetworkingConfig().getMaxPayloadSize(),
-                  config.isRpcSnappyAircompressorEnabled());
+              RpcEncoding.createSszSnappyEncoding(spec.getNetworkingConfig().getMaxPayloadSize());
         }
         final UInt256 discoveryNodeId = DISCOVERY_NODE_ID_GENERATOR.next();
         final int numberOfColumns = spec.getNumberOfDataColumns().orElse(0);